move working resource pool business into ResourceSelection
