+++ kolibri/plugins/coach/assets/src/views/plan/CreateExamPage/CreateQuizSection.vue
@@ -394,7 +394,6 @@
         allSections,
         activeSection,
         inactiveSections,
-        workingResourcePool,
         activeQuestionsPool,
         activeResourceMap,
         activeResourcePool,
@@ -445,7 +444,6 @@
         allSections,
         activeSection,
         inactiveSections,
-        workingResourcePool,
         activeResourceMap,
         activeResourcePool,
         activeQuestionsPool,
